+++ scripts/test/Selenium/Agent/AgentTicketMove.t
@@ -237,6 +237,31 @@ $Selenium->RunTest(
             "ACL restriction error message found for 'Spam' menu",
         );
 
+        # Test for bug#12559 that nothing shpuld happen, if the user click on a disabled queue (only for move type 'form').
+        $Helper->ConfigSettingChange(
+            Valid => 1,
+            Key   => 'Ticket::Frontend::MoveType',
+            Value => 'form'
+        );
+
+        # Reload the page, to get the new sys config option.
+        $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AgentTicketZoom;TicketID=$TicketID");
+        $Selenium->execute_script("\$('#DestQueueID').val('4').trigger('redraw.InputField').trigger('change');");
+
+        # Check that nothing happens, after the queue selection in the dropdown.
+        $Self->True(
+            index( $Selenium->get_current_url(), "${ScriptAlias}index.pl?Action=AgentTicketZoom;TicketID=$TicketID" ) > -1,
+            'The current url is the same (no reload happens).',
+        );
+
+        $Self->Is(
+            $Selenium->execute_script(
+                "return \$('p.Value[title=\"Misc\"]').text()"
+            ),
+            'Misc',
+            'The Queue was not changed.',
+        );
+
         # delete test ACL
         my $Success = $ACLObject->ACLDelete(
             ID     => $ACLID,